refactoring welcome files, registering partially works
