+++ frontend/src/locales/de/translation.json
@@ -1,60 +1,64 @@
 {
-	"Your portal gun must be busted ...": "Dein Portalstrahler muss kaputt sein...",
-	"Your curious ass will be redirected in": "Dein neugieriger Hintern wird umgeleitet in",
-	"Join": "Beitreten",
-	"Password": "Passwort",
-	"Email": "E-Mail",
-	"Sign In via Email": "Anmelden per E-Mail",
-	"Sign In via 42": "Anmelden über 42",
-	"Friends": "Freunde",
-	"Language": "Sprache",
-	"Original Pong": "Original Pong",
-	"Multiplayer": "Mehrspieler",
-	"Modded Pong": "Modifiziertes Pong",
-	"3D Pong": "3D Pong",
-	"Pong against AI": "Pong gegen KI",
-	"Online": "Online",
-	"Channels": "Kanäle",
-	"General": "Allgemein",
-	"Random": "Zufällig",
-	"Back": "Zurück",
-	"Type": "Typ",
-	"Type your message here...": "Tippen Sie Ihre Nachricht hier ein...",
-	"Choose File": "Datei auswählen",
-	"Name": "Name",
-	"Intra": "Intra",
-	"Match History": "Spielverlauf",
-	"Opponent": "Gegner",
-	"Score": "Punktzahl",
-	"About Me": "Über mich",
-	"User Details": "Benutzerdetails",
-	"Age": "Alter",
-	"years old": "Jahre alt",
-	"Gender": "Geschlecht",
-	"School": "Schule",
-	"Level": "Level",
-	"Win Rate": "Gewinnrate",
-	"Result": "Ergebnis",
-	"Win": "Gewinnen",
-	"Loss": "Verlust",
-	"Game": "Spiel",
-	"Pong": "Pong",
-	"Original": "Original",
-	"Modded": "Modifiziert",
-	"3D": "3D",
-	"Welcome To": "Willkommen bei",
-	"TranscENDence": "TranscENDence",
-	"Where Your Core Journey": "Wo Ihre Kernreise",
-	"Ends": "endet",
-	"Home": "Startseite",
-	"Channels & Private Messages": "Kanäle & Private Nachrichten",
-	"Play or Watch Games": "Spiele spielen oder zuschauen",
-	"Profile": "Profil",
-	"About Us": "Über uns",
-	"Logout": "Abmelden",
-	"About zstenger": "Der Versionskontrollvirtuose, der sich durch das Chaos navigiert. Bestätigt Frühstückswahlen, erstellt Wochenendpläne und verschmilzt Lebensentscheidungen einwandfrei. Sein Code ist so organisiert wie sein Repository, und 'git push' ist sein Mantra. Die Legende besagt, dass er sogar seine Träume verzweigt.",
-	"About jergashe": "Wissenshungrig, aber lässt gelegentlich die Datenbanktabelle fallen auf der Suche nach Verständnis. Der ewige Lernende, der es liebt, Mysterien zu entwirren, selbst wenn es einen 'DROP TABLE'-Umweg bedeutet.",
-	"About asioud": "Ein Maestro der Automatisierung und Bereitstellungsorchestrierung, der jedoch gelegentlich am Rande des Chaos tanzt. Die Balance zwischen Genialität und 'Oops', dieser DevOps-Guru verwandelt jede Niederlage in eine wertvolle Lektion. Bereitstellungsskripte tagsüber, Debugging-Beichten nachts.",
-	"About kvebers": "Meister des kodierten Chaos, genährt von Burst von Hyperfokus. Multitasking-Maestro mit einem Gehirn schneller als Kompilierzeiten. Springt zwischen Projekten wie ein koffeinierter Eichhörnchen auf einer Tastatur. Bugs, nehmt euch in Acht, er wird euch in die Debugging-Vergessenheit bringen.",
-	"About slaszlo": "Die Anpassung ist sein zweiter Vorname, und er navigiert durch das Leben mit der Präzision eines gut ausgearbeiteten Skripts. Eigenständige Software? Nein danke. Er lebt für das süße Gefühl von Freiheit und Zusammenarbeit in der Linux-Welt. Tux der Pinguin ist sein spirituelles Tier, und das Terminal ist sein glücklicher Ort."
+  "Sign In": "Anmelden",
+  "Login": "Anmeldung",
+  "Register": "Registrieren",
+  "Username": "Benutzername",
+  "Your portal gun must be busted ...": "Dein Portalstrahler muss kaputt sein...",
+  "Your curious ass will be redirected in": "Dein neugieriger Hintern wird umgeleitet in",
+  "Join": "Beitreten",
+  "Password": "Passwort",
+  "Email": "E-Mail",
+  "Sign In via Email": "Anmelden per E-Mail",
+  "Sign In via 42": "Anmelden über 42",
+  "Friends": "Freunde",
+  "Language": "Sprache",
+  "Original Pong": "Original Pong",
+  "Multiplayer": "Mehrspieler",
+  "Modded Pong": "Modifiziertes Pong",
+  "3D Pong": "3D Pong",
+  "Pong against AI": "Pong gegen KI",
+  "Online": "Online",
+  "Channels": "Kanäle",
+  "General": "Allgemein",
+  "Random": "Zufällig",
+  "Back": "Zurück",
+  "Type": "Typ",
+  "Type your message here...": "Tippen Sie Ihre Nachricht hier ein...",
+  "Choose File": "Datei auswählen",
+  "Name": "Name",
+  "Intra": "Intra",
+  "Match History": "Spielverlauf",
+  "Opponent": "Gegner",
+  "Score": "Punktzahl",
+  "About Me": "Über mich",
+  "User Details": "Benutzerdetails",
+  "Age": "Alter",
+  "years old": "Jahre alt",
+  "Gender": "Geschlecht",
+  "School": "Schule",
+  "Level": "Level",
+  "Win Rate": "Gewinnrate",
+  "Result": "Ergebnis",
+  "Win": "Gewinnen",
+  "Loss": "Verlust",
+  "Game": "Spiel",
+  "Pong": "Pong",
+  "Original": "Original",
+  "Modded": "Modifiziert",
+  "3D": "3D",
+  "Welcome To": "Willkommen bei",
+  "TranscENDence": "TranscENDence",
+  "Where Your Core Journey": "Wo Ihre Kernreise",
+  "Ends": "endet",
+  "Home": "Startseite",
+  "Channels & Private Messages": "Kanäle & Private Nachrichten",
+  "Play or Watch Games": "Spiele spielen oder zuschauen",
+  "Profile": "Profil",
+  "About Us": "Über uns",
+  "Logout": "Abmelden",
+  "About zstenger": "Der Versionskontrollvirtuose, der sich durch das Chaos navigiert. Bestätigt Frühstückswahlen, erstellt Wochenendpläne und verschmilzt Lebensentscheidungen einwandfrei. Sein Code ist so organisiert wie sein Repository, und 'git push' ist sein Mantra. Die Legende besagt, dass er sogar seine Träume verzweigt.",
+  "About jergashe": "Wissenshungrig, aber lässt gelegentlich die Datenbanktabelle fallen auf der Suche nach Verständnis. Der ewige Lernende, der es liebt, Mysterien zu entwirren, selbst wenn es einen 'DROP TABLE'-Umweg bedeutet.",
+  "About asioud": "Ein Maestro der Automatisierung und Bereitstellungsorchestrierung, der jedoch gelegentlich am Rande des Chaos tanzt. Die Balance zwischen Genialität und 'Oops', dieser DevOps-Guru verwandelt jede Niederlage in eine wertvolle Lektion. Bereitstellungsskripte tagsüber, Debugging-Beichten nachts.",
+  "About kvebers": "Meister des kodierten Chaos, genährt von Burst von Hyperfokus. Multitasking-Maestro mit einem Gehirn schneller als Kompilierzeiten. Springt zwischen Projekten wie ein koffeinierter Eichhörnchen auf einer Tastatur. Bugs, nehmt euch in Acht, er wird euch in die Debugging-Vergessenheit bringen.",
+  "About slaszlo": "Die Anpassung ist sein zweiter Vorname, und er navigiert durch das Leben mit der Präzision eines gut ausgearbeiteten Skripts. Eigenständige Software? Nein danke. Er lebt für das süße Gefühl von Freiheit und Zusammenarbeit in der Linux-Welt. Tux der Pinguin ist sein spirituelles Tier, und das Terminal ist sein glücklicher Ort."
 }
