+++ frontend/src/locales/en/translation.json
@@ -1,4 +1,8 @@
 {
+  "Sign In": "Sign In",
+  "Login": "Login",
+  "Register": "Register",
+  "Username": "Username",
   "Your portal gun must be busted ...": "Your portal gun must be busted ...",
   "Your curious ass will be redirected in": "Your curious ass will be redirected in",
   "Join": "Join",
