+++ frontend/src/locales/fr/translation.json
@@ -1,60 +1,64 @@
 {
-	"Your portal gun must be busted ...": "Votre pistolet à portails doit être défectueux...",
-	"Your curious ass will be redirected in": "Votre cul curieux sera redirigé dans",
-	"Join": "Rejoindre",
-	"Password": "Mot de passe",
-	"Email": "E-mail",
-	"Sign In via Email": "Se connecter via e-mail",
-	"Sign In via 42": "Se connecter via 42",
-	"Friends": "Amis",
-	"Language": "Langue",
-	"Original Pong": "Pong Original",
-	"Multiplayer": "Multijoueur",
-	"Modded Pong": "Pong modifié",
-	"3D Pong": "Pong 3D",
-	"Pong against AI": "Pong contre l'IA",
-	"Online": "En ligne",
-	"Channels": "Canaux",
-	"General": "Général",
-	"Random": "Aléatoire",
-	"Back": "Retour",
-	"Type": "Type",
-	"Type your message here...": "Tapez votre message ici...",
-	"Choose File": "Choisir un fichier",
-	"Name": "Nom",
-	"Intra": "Intra",
-	"Match History": "Historique des matchs",
-	"Opponent": "Adversaire",
-	"Score": "Score",
-	"About Me": "À propos de moi",
-	"User Details": "Détails de l'utilisateur",
-	"Age": "Âge",
-	"years old": "ans",
-	"Gender": "Genre",
-	"School": "École",
-	"Level": "Niveau",
-	"Win Rate": "Taux de victoire",
-	"Result": "Résultat",
-	"Win": "Gagner",
-	"Loss": "Perte",
-	"Game": "Jeu",
-	"Pong": "Pong",
-	"Original": "Original",
-	"Modded": "Modifié",
-	"3D": "3D",
-	"Welcome To": "Bienvenue à",
-	"TranscENDence": "TranscENDence",
-	"Where Your Core Journey": "Où se termine votre voyage central",
-	"Ends": "Se termine",
-	"Home": "Accueil",
-	"Channels & Private Messages": "Canaux et messages privés",
-	"Play or Watch Games": "Jouer ou regarder des jeux",
-	"Profile": "Profil",
-	"About Us": "À propos de nous",
-	"Logout": "Déconnexion",
-	"About zstenger": "Le virtuose du contrôle de version, naviguant à travers le chaos. Valide les choix de petit-déjeuner, crée des branches pour les plans du week-end et fusionne les décisions de vie sans faille. Son code est aussi organisé que son dépôt, et 'git push' est son mantra. La légende dit même qu'il bifurque ses rêves.",
-	"About jergashe": "Affamé de connaissances, mais laisse parfois tomber la table de la base de données dans sa quête de compréhension. L'apprenant éternel qui aime démêler les mystères, même si cela signifie un détour par 'DROP TABLE'.",
-	"About asioud": "Maestro de l'automatisation et de l'orchestration du déploiement, mais danse parfois sur le bord du chaos. Équilibrant entre la brillance et les 'oops', ce gourou DevOps transforme chaque revers en une leçon précieuse. Scripts de déploiement le jour, confessions de débogage la nuit.",
-	"About kvebers": "Maître du chaos de codage, alimenté par des rafales d'hyperconcentration. Maestro multitâche avec un cerveau plus rapide que les temps de compilation. Sautant entre les projets comme un écureuil caféiné sur un clavier. Les bugs prenez garde, il vous débuguera dans l'oubli.",
-	"About slaszlo": "La personnalisation est son deuxième prénom, et il navigue dans la vie avec la précision d'un script bien conçu. Logiciel propriétaire ? Non merci. Il vit pour cette douce sensation de liberté et de collaboration dans le royaume Linux. Tux le pingouin est son animal spirituel, et le terminal est son lieu heureux."
+  "Sign In": "Se connecter",
+  "Login": "Connexion",
+  "Register": "S'inscrire",
+  "Username": "Nom d'utilisateur",
+  "Your portal gun must be busted ...": "Votre pistolet à portails doit être défectueux...",
+  "Your curious ass will be redirected in": "Votre cul curieux sera redirigé dans",
+  "Join": "Rejoindre",
+  "Password": "Mot de passe",
+  "Email": "E-mail",
+  "Sign In via Email": "Se connecter via e-mail",
+  "Sign In via 42": "Se connecter via 42",
+  "Friends": "Amis",
+  "Language": "Langue",
+  "Original Pong": "Pong Original",
+  "Multiplayer": "Multijoueur",
+  "Modded Pong": "Pong modifié",
+  "3D Pong": "Pong 3D",
+  "Pong against AI": "Pong contre l'IA",
+  "Online": "En ligne",
+  "Channels": "Canaux",
+  "General": "Général",
+  "Random": "Aléatoire",
+  "Back": "Retour",
+  "Type": "Type",
+  "Type your message here...": "Tapez votre message ici...",
+  "Choose File": "Choisir un fichier",
+  "Name": "Nom",
+  "Intra": "Intra",
+  "Match History": "Historique des matchs",
+  "Opponent": "Adversaire",
+  "Score": "Score",
+  "About Me": "À propos de moi",
+  "User Details": "Détails de l'utilisateur",
+  "Age": "Âge",
+  "years old": "ans",
+  "Gender": "Genre",
+  "School": "École",
+  "Level": "Niveau",
+  "Win Rate": "Taux de victoire",
+  "Result": "Résultat",
+  "Win": "Gagner",
+  "Loss": "Perte",
+  "Game": "Jeu",
+  "Pong": "Pong",
+  "Original": "Original",
+  "Modded": "Modifié",
+  "3D": "3D",
+  "Welcome To": "Bienvenue à",
+  "TranscENDence": "TranscENDence",
+  "Where Your Core Journey": "Où se termine votre voyage central",
+  "Ends": "Se termine",
+  "Home": "Accueil",
+  "Channels & Private Messages": "Canaux et messages privés",
+  "Play or Watch Games": "Jouer ou regarder des jeux",
+  "Profile": "Profil",
+  "About Us": "À propos de nous",
+  "Logout": "Déconnexion",
+  "About zstenger": "Le virtuose du contrôle de version, naviguant à travers le chaos. Valide les choix de petit-déjeuner, crée des branches pour les plans du week-end et fusionne les décisions de vie sans faille. Son code est aussi organisé que son dépôt, et 'git push' est son mantra. La légende dit même qu'il bifurque ses rêves.",
+  "About jergashe": "Affamé de connaissances, mais laisse parfois tomber la table de la base de données dans sa quête de compréhension. L'apprenant éternel qui aime démêler les mystères, même si cela signifie un détour par 'DROP TABLE'.",
+  "About asioud": "Maestro de l'automatisation et de l'orchestration du déploiement, mais danse parfois sur le bord du chaos. Équilibrant entre la brillance et les 'oops', ce gourou DevOps transforme chaque revers en une leçon précieuse. Scripts de déploiement le jour, confessions de débogage la nuit.",
+  "About kvebers": "Maître du chaos de codage, alimenté par des rafales d'hyperconcentration. Maestro multitâche avec un cerveau plus rapide que les temps de compilation. Sautant entre les projets comme un écureuil caféiné sur un clavier. Les bugs prenez garde, il vous débuguera dans l'oubli.",
+  "About slaszlo": "La personnalisation est son deuxième prénom, et il navigue dans la vie avec la précision d'un script bien conçu. Logiciel propriétaire ? Non merci. Il vit pour cette douce sensation de liberté et de collaboration dans le royaume Linux. Tux le pingouin est son animal spirituel, et le terminal est son lieu heureux."
 }
