+++ frontend/src/locales/hu/translation.json
@@ -1,55 +1,59 @@
 {
-	"Your portal gun must be busted ...": "Valószínűleg a portálpisztolyod elbaszódott ...",
-	"Your curious ass will be redirected in": "A kíváncsi feneked át lesz irányítva",
-	"Join": "Csatlakozás",
-	"Password": "Jelszó",
-	"Email": "E-mail",
-	"Sign In via Email": "E-mail Bejelentkezés",
-	"Sign In via 42": "42 Bejelentkezés",
-	"Friends": "Barátok",
-	"Language": "Nyelv",
-	"Original Pong": "Eredeti Pong",
-	"Multiplayer": "Többjátékos",
-	"Modded Pong": "Módosított Pong",
-	"3D Pong": "3D Pong",
-	"Pong against AI": "Pong mesterséges intelligencia ellen",
-	"Online": "Elérhető",
-	"Channels": "Csatornák",
-	"General": "Általános",
-	"Random": "Véletlenszerű",
-	"Back": "Vissza",
-	"Type": "Típus",
-	"Type your message here...": "Írja be üzenetét itt...",
-	"Choose File": "Fájl kiválasztása",
-	"Name": "Név",
-	"Intra": "Intra",
-	"Match History": "Mérkőzések előzményei",
-	"Opponent": "Ellenfél",
-	"Score": "Pontszám",
-	"About Me": "Rólam",
-	"User Details": "Felhasználó adatai",
-	"Age": "Kor",
-	"years old": "éves",
-	"Gender": "Nem",
-	"School": "Iskola",
-	"Level": "Szint",
-	"Win Rate": "Győzelmi arány",
-	"Result": "Eredmény",
-	"Win": "Győzelem",
-	"Loss": "Veszteség",
-	"Game": "Játék",
-	"Pong": "Pong",
-	"Original": "Eredeti",
-	"Modded": "Módosított",
-	"3D": "3D",
-	"Welcome To": "Üdvözlünk a",
-	"TranscENDence": "TranscENDencenél",
-	"Where Your Core Journey": "Ahol az utad",
-	"Ends": "Véget ér",
-	"Home": "Kezdőlap",
-	"Channels & Private Messages": "Csatornák és privát üzenetek",
-	"Play or Watch Games": "Játssz vagy nézz játékokat",
-	"Profile": "Profil",
-	"About Us": "Rólunk",
-	"Logout": "Kijelentkezés"
+  "Sign In": "Bejelentkezés",
+  "Login": "Bejelentkezés",
+  "Register": "Regisztráció",
+  "Username": "Felhasználónév",
+  "Your portal gun must be busted ...": "Valószínűleg a portálpisztolyod elbaszódott ...",
+  "Your curious ass will be redirected in": "A kíváncsi feneked át lesz irányítva",
+  "Join": "Csatlakozás",
+  "Password": "Jelszó",
+  "Email": "E-mail",
+  "Sign In via Email": "E-mail Bejelentkezés",
+  "Sign In via 42": "42 Bejelentkezés",
+  "Friends": "Barátok",
+  "Language": "Nyelv",
+  "Original Pong": "Eredeti Pong",
+  "Multiplayer": "Többjátékos",
+  "Modded Pong": "Módosított Pong",
+  "3D Pong": "3D Pong",
+  "Pong against AI": "Pong mesterséges intelligencia ellen",
+  "Online": "Elérhető",
+  "Channels": "Csatornák",
+  "General": "Általános",
+  "Random": "Véletlenszerű",
+  "Back": "Vissza",
+  "Type": "Típus",
+  "Type your message here...": "Írja be üzenetét itt...",
+  "Choose File": "Fájl kiválasztása",
+  "Name": "Név",
+  "Intra": "Intra",
+  "Match History": "Mérkőzések előzményei",
+  "Opponent": "Ellenfél",
+  "Score": "Pontszám",
+  "About Me": "Rólam",
+  "User Details": "Felhasználó adatai",
+  "Age": "Kor",
+  "years old": "éves",
+  "Gender": "Nem",
+  "School": "Iskola",
+  "Level": "Szint",
+  "Win Rate": "Győzelmi arány",
+  "Result": "Eredmény",
+  "Win": "Győzelem",
+  "Loss": "Veszteség",
+  "Game": "Játék",
+  "Pong": "Pong",
+  "Original": "Eredeti",
+  "Modded": "Módosított",
+  "3D": "3D",
+  "Welcome To": "Üdvözlünk a",
+  "TranscENDence": "TranscENDencenél",
+  "Where Your Core Journey": "Ahol az utad",
+  "Ends": "Véget ér",
+  "Home": "Kezdőlap",
+  "Channels & Private Messages": "Csatornák és privát üzenetek",
+  "Play or Watch Games": "Játssz vagy nézz játékokat",
+  "Profile": "Profil",
+  "About Us": "Rólunk",
+  "Logout": "Kijelentkezés"
 }
