+++ frontend/src/locales/jp/translation.json
@@ -1,60 +1,64 @@
 {
-	"Your portal gun must be busted ...": "おそらくあなたのポータルガンが故障しています...",
-	"Your curious ass will be redirected in": "あなたの好奇心旺盛なお尻は、あとでリダイレクトされます",
-	"Join": "参加する",
-	"Password": "パスワード",
-	"Email": "メール",
-	"Sign In via Email": "メールでサインイン",
-	"Sign In via 42": "42経由でサインイン",
-	"Friends": "友達",
-	"Language": "言語",
-	"Original Pong": "オリジナルポン",
-	"Multiplayer": "マルチプレイヤー",
-	"Modded Pong": "変更済みポン",
-	"3D Pong": "3Dポン",
-	"Pong against AI": "AIに対するポン",
-	"Online": "オンライン",
-	"Channels": "チャンネル",
-	"General": "一般",
-	"Random": "ランダム",
-	"Back": "戻る",
-	"Type": "タイプ",
-	"Type your message here...": "ここにメッセージを入力してください...",
-	"Choose File": "ファイルを選択",
-	"Name": "名前",
-	"Intra": "インタ",
-	"Match History": "対戦履歴",
-	"Opponent": "相手",
-	"Score": "スコア",
-	"About Me": "自己紹介",
-	"User Details": "ユーザー詳細",
-	"Age": "年齢",
-	"years old": "歳",
-	"Gender": "性別",
-	"School": "学校",
-	"Level": "レベル",
-	"Win Rate": "勝率",
-	"Result": "結果",
-	"Win": "勝利",
-	"Loss": "敗北",
-	"Game": "ゲーム",
-	"Pong": "ポン",
-	"Original": "オリジナル",
-	"Modded": "変更済み",
-	"3D": "3D",
-	"Welcome To": "ようこそ",
-	"TranscENDence": "トランセンデンス",
-	"Where Your Core Journey": "あなたのコアな旅が始まる場所",
-	"Ends": "終了",
-	"Home": "ホーム",
-	"Channels & Private Messages": "チャンネルとプライベートメッセージ",
-	"Play or Watch Games": "ゲームをプレイするか見るか",
-	"Profile": "プロフィール",
-	"About Us": "私たちについて",
-	"Logout": "ログアウト",
-	"About zstenger": "バージョン管理のバーチュオーゾ、カオスを航海します。朝食の選択を確定し、週末の計画のためにブランチを作成し、人生の決断を完璧に統合します。彼のコードは彼のリポジトリと同じくらい整理されており、「git push」は彼のマントラです。伝説によれば、彼は夢すら分岐させると言います。",
-	"About jergashe": "知識に飢えていますが、理解を求めてデータベーステーブルを時折ドロップします。永遠の学習者は、たとえそれが 'DROP TABLE' の迂回路を意味する場合でも、謎を解き明かすのが好きです。",
-	"About asioud": "自動化とデプロイオーケストレーションのマエストロですが、時折混沌の縁で踊ります。輝きと 'oops' の微妙な線のバランスを取るこのDevOpsグルは、すべての逆境を貴重な教訓に変えます。デプロイメントスクリプトは昼間、デバッグの告白は夜に。",
-	"About kvebers": "コーディングの混沌のマスター、ハイパーフォーカスのバーストで駆り立てられています。コンパイルタイムよりも速い脳を持つマルチタスキングのマエストロ。キーボード上のカフェイン摂取リスケーティングスクイレルのようにプロジェクト間を移動します。バグよ注意、あなたを無意識の中にデバッグします。",
-	"About slaszlo": "カスタマイズは彼のミドルネームであり、彼はウェルクラフトなスクリプトの精密さで人生を航海します。プロプライエタリソフトウェア？いいえ、ありがとう。彼はLinuxの領域での自由と協力のあの甘い感覚のために生きています。Tux the penguinは彼のスピリットアニマルであり、端末は彼のハッピープレイスです。"
+  "Sign In": "サインイン",
+  "Login": "ログイン",
+  "Register": "登録",
+  "Username": "ユーザー名",
+  "Your portal gun must be busted ...": "おそらくあなたのポータルガンが故障しています...",
+  "Your curious ass will be redirected in": "あなたの好奇心旺盛なお尻は、あとでリダイレクトされます",
+  "Join": "参加する",
+  "Password": "パスワード",
+  "Email": "メール",
+  "Sign In via Email": "メールでサインイン",
+  "Sign In via 42": "42経由でサインイン",
+  "Friends": "友達",
+  "Language": "言語",
+  "Original Pong": "オリジナルポン",
+  "Multiplayer": "マルチプレイヤー",
+  "Modded Pong": "変更済みポン",
+  "3D Pong": "3Dポン",
+  "Pong against AI": "AIに対するポン",
+  "Online": "オンライン",
+  "Channels": "チャンネル",
+  "General": "一般",
+  "Random": "ランダム",
+  "Back": "戻る",
+  "Type": "タイプ",
+  "Type your message here...": "ここにメッセージを入力してください...",
+  "Choose File": "ファイルを選択",
+  "Name": "名前",
+  "Intra": "インタ",
+  "Match History": "対戦履歴",
+  "Opponent": "相手",
+  "Score": "スコア",
+  "About Me": "自己紹介",
+  "User Details": "ユーザー詳細",
+  "Age": "年齢",
+  "years old": "歳",
+  "Gender": "性別",
+  "School": "学校",
+  "Level": "レベル",
+  "Win Rate": "勝率",
+  "Result": "結果",
+  "Win": "勝利",
+  "Loss": "敗北",
+  "Game": "ゲーム",
+  "Pong": "ポン",
+  "Original": "オリジナル",
+  "Modded": "変更済み",
+  "3D": "3D",
+  "Welcome To": "ようこそ",
+  "TranscENDence": "トランセンデンス",
+  "Where Your Core Journey": "あなたのコアな旅が始まる場所",
+  "Ends": "終了",
+  "Home": "ホーム",
+  "Channels & Private Messages": "チャンネルとプライベートメッセージ",
+  "Play or Watch Games": "ゲームをプレイするか見るか",
+  "Profile": "プロフィール",
+  "About Us": "私たちについて",
+  "Logout": "ログアウト",
+  "About zstenger": "バージョン管理のバーチュオーゾ、カオスを航海します。朝食の選択を確定し、週末の計画のためにブランチを作成し、人生の決断を完璧に統合します。彼のコードは彼のリポジトリと同じくらい整理されており、「git push」は彼のマントラです。伝説によれば、彼は夢すら分岐させると言います。",
+  "About jergashe": "知識に飢えていますが、理解を求めてデータベーステーブルを時折ドロップします。永遠の学習者は、たとえそれが 'DROP TABLE' の迂回路を意味する場合でも、謎を解き明かすのが好きです。",
+  "About asioud": "自動化とデプロイオーケストレーションのマエストロですが、時折混沌の縁で踊ります。輝きと 'oops' の微妙な線のバランスを取るこのDevOpsグルは、すべての逆境を貴重な教訓に変えます。デプロイメントスクリプトは昼間、デバッグの告白は夜に。",
+  "About kvebers": "コーディングの混沌のマスター、ハイパーフォーカスのバーストで駆り立てられています。コンパイルタイムよりも速い脳を持つマルチタスキングのマエストロ。キーボード上のカフェイン摂取リスケーティングスクイレルのようにプロジェクト間を移動します。バグよ注意、あなたを無意識の中にデバッグします。",
+  "About slaszlo": "カスタマイズは彼のミドルネームであり、彼はウェルクラフトなスクリプトの精密さで人生を航海します。プロプライエタリソフトウェア？いいえ、ありがとう。彼はLinuxの領域での自由と協力のあの甘い感覚のために生きています。Tux the penguinは彼のスピリットアニマルであり、端末は彼のハッピープレイスです。"
 }
