+++ frontend/src/pages/Welcome.js
@@ -1,0 +1,41 @@
+import React, { useState } from "react";
+import { useNavigate } from "react-router-dom";
+import { useTranslation } from "react-i18next";
+import LanguageButton from "../components/welcome_page/LanguageButton";
+import SignInButton from "../components/welcome_page/SignInButton";
+import RegisterButton from "../components/welcome_page/RegisterButton";
+import LoginVia42 from "../components/welcome_page/LoginVia42";
+
+
+
+function Welcome() {
+  const { t } = useTranslation();
+  const navigate = useNavigate();
+  const [email] = useState("");
+  const [password] = useState("");
+
+  const redirectToHome = (isJoinButton) => {
+    if (isJoinButton && (!email || !password)) {
+      alert("Please enter email and password");
+    } else {
+      navigate("/home");
+    }
+  };
+
+  return (
+    <div
+      className="flex flex-col items-center justify-center 
+      h-screen bg-cover bg-center bg-no-repeat"
+    >
+      <LanguageButton />
+      <div className="flex flex-col items-center justify-center flex-grow">
+	  	<SignInButton t={t} redirectToHome={redirectToHome} />
+	  	<RegisterButton t={t} redirectToHome={redirectToHome} />
+		<LoginVia42 t={t} navigate={navigate} />
+      </div>
+      <div className="flex items-center justify-center"></div>
+    </div>
+  );
+}
+
+export default Welcome;
